마카오뱅크 송금3

+++ week07/makaoBankTransfer/20220803-조성환/app/src/main/java/models/Account.java
@@ -24,7 +24,7 @@ public class Account {
     return amount;
   }
 
-  public void transfer(Account receiver, int amount) {
+  public void transfer(Account receiver, long amount) {
     this.amount -= amount;
 
     receiver.amount += amount;
