+++ week07/makaoBankTransfer/20220803-조성환/app/src/main/java/services/TransferService.java
@@ -13,7 +13,7 @@ public class TransferService {
     this.accounts = new ArrayList<>(accounts);
   }
 
-  public void transfer(String from, String to, int amount) {
+  public void transfer(String from, String to, long amount) {
     Account sender = accounts.stream()
         .filter(account -> account.identifier().equals(from))
         .findFirst().get();
